fix: dev script after inquirer update
