+++ package.json
@@ -60,7 +60,6 @@
     "glob": "^10.4.3",
     "http-server": "14.1.1",
     "husky": "9.0.11",
-    "inquirer": "^10.0.0",
     "jscpd": "^4.0.5",
     "lint-staged": "^15.2.7",
     "markdownlint-cli": "^0.41.0",
