+++ scripts/package.json
@@ -23,7 +23,7 @@
     "build-showcases:vue": "npm run build --workspace=vue-showcase",
     "build:components": "npm run build --workspace=@db-ui/components",
     "build:foundations": "npm run build --workspace=@db-ui/foundations",
-    "dev": "node dev.js",
+    "dev": "tsx dev.js",
     "dev:angular-components": "npm run dev:angular --workspace=@db-ui/components",
     "dev:plain-html-components": "npm run dev:html --workspace=@db-ui/components",
     "dev:react-components": "npm run dev:react --workspace=@db-ui/components",
@@ -63,7 +63,9 @@
     "update:icon-fonts": "npm-run-all generate:icon-fonts generate:icon-types"
   },
   "devDependencies": {
+    "@inquirer/prompts": "5.1.2",
     "glob": "^10.4.3",
+    "inquirer": "^10.0.0",
     "replace-in-file": "^8.1.0"
   }
 }
